Fix EXTR for 32bit
This fixes some minor things in shifti and handle_add (for ADD extended
reg, which uses SP[] for input always, also when setflags).  And it
fixes one major bug in handle_extr, where it wasn't using the extended
input, but the input reg itself for shifting, leading to leaking bits
in ROR immediate (an alias of EXTR), which in turn leads to problems
in crypt code (here SHA1).  Now libgcrypt works.