Don't truncate paths to less than PATH_MAX
resolve_dirfd_path used a buffer of 1024 (< PATH_MAX) and readlink
doesn't return errors when the buffer is too small (it silently
truncates), so when the path was very long (> 1024) it got truncated
and wrong files or error codes were returned.  Using PATH_MAX
as buffer size readlink won't truncate (if the name is larger it will
correctly return ENAMETOOLONG, and the callers will do the right thing).